requirements.txt
azure-identity==1.25.2
openai==2.26.0
agent-framework-azure-ai==1.0.0rc3
agent-framework-core==1.0.0rc3
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-agentserver-core==1.0.0b16
gremlinpython>=3.7.0
azure-cosmos>=4.7.0
requests
pandas
flask
pypdfium2
pillow
python-dotenv